GS data analysis values first vertical offset and then vertical scaled.. Barometer altitude NaN problem fixed, kalman matrix value order fixed (F matrix)

+++ Sidus_Github_Solution/Drone_Flight_Controller_2/Config.h
@@ -249,7 +249,7 @@ float ypr[3];           // [yaw, pitch, roll]   yaw/pitch/roll container and gra
 
 
 float barometerTemp = 0.0;
-float barometerAlt = 0.0;
+double barometerAlt = 0.0;
 int32_t barometerPress;
 
 float compassHdg;
